Fix FDX/FLAT sort order and 64-bit architecture packing

+++ fdx.h
@@ -12,7 +12,7 @@ struct fdx_entry {
 };
 
 struct fdx_entries {
-  size_t numEntries;
+  uint32_t numEntries;
   struct fdx_entry ** entries;
 };
